+++ tests/PrimitiveTypesTest.php
@@ -3,30 +3,18 @@ declare(strict_types=1);
 namespace codename\parquet\tests;
 
 use codename\parquet\data\Schema;
+use codename\parquet\data\DataType;
 use codename\parquet\data\DataField;
 use codename\parquet\data\DataColumn;
 
 final class PrimitiveTypesTest extends TestBase
 {
   /**
-   * [testWriteLoadsOfBooleansAllTrue description]
-   */
-  public function testWriteLoadsOfBooleansAllTrue(): void {
-    $counts = [
-      100,
-      1000
-    ];
-
-    foreach($counts as $count) {
-      $this->Write_loads_of_booleans_all_true($count);
-    }
-  }
-
-  /**
-   * [Write_loads_of_booleans_all_true description]
+   * @testWith [100]
+   *           [1000]
    * @param int $count [description]
    */
-  public function Write_loads_of_booleans_all_true(int $count):void
+  public function testWriteLoadsOfBooleansAllTrue(int $count):void
   {
     $id = DataField::createFromType("enabled", 'boolean');
     $schema = new Schema([$id]);
@@ -34,10 +22,36 @@ final class PrimitiveTypesTest extends TestBase
     $data = array_fill(0, $count, true);
 
     $read = $this->WriteReadSingleColumn($id, new DataColumn($id, $data));
-    
+
     for($i = 0; $i < $count; $i++)
     {
       $this->assertTrue($read->getData()[$i], "got FALSE at position {$i}");
+    }
+  }
+
+  /**
+   * Maximum value for a 32-bit unsigned integer
+   * @var int
+   */
+  const UIntmaxValue = 4294967295;
+
+  /**
+   * @testWith [100]
+   * @param int $count  [description]
+   */
+  public function testWriteBunchOfUInts(int $count): void {
+
+    $id = new DataField('id', DataType::UnsignedInt32, false);
+
+    $data = [];
+    for ($i=0; $i < $count; $i++) {
+      $data[$i] = static::UIntmaxValue - $i;
+    }
+
+    $read = $this->WriteReadSingleColumn($id, new DataColumn($id, $data));
+    for ($i=0; $i < $count; $i++) {
+      $result = $read->getData()[$i];
+      $this->assertEquals(static::UIntmaxValue - $i, $result);
     }
   }
 }